C030_Matias_MoreThanThis
Recently tensions have risen between Greece and Altis due to land disputes. The Greek government has reason to believed that the Altian government has been developing a WMD in the island of Stratis and that it has recently been brought over to Altis.
Contractors have been hired by the greek government to infiltrate the port on which it is believed that the weapon was unloaded and find any proof that the weapon was there. Should they find proof they are to move in on a research facility which is believed to be housing the weapon and destroy it. Communications at the facility have been sabotaged by a greek operative but they failed to do so at the port, so if contractors are detected while infiltrating said port it is expected that reinforcements will be called on the facility.
Once the objectives are completed contractors are to meet with a CTRG team that is waiting to extract them out of the country. In order to not leave any trace of our pressence contractors have been provided with civilian vehicles and must retrieve any gear that is used during the operation.
            <br/>
            <br/>Contractors have been provided with an SDV and wetsuits in order to covertly infiltrate the port. This SDV needs to be destroyed before contractors leave the area as it would be tangible proof of Greek involvenment.